Changed interprocess communication from socket.Socket to a multiproccessing.connection Listener/Client model.
Also we are now sending the whole message as one long packet. This was a
compromise of sending byte for byte in old method for speed.  The old
method was better emulation of actual DotStar LED modules. But this method
is faster, which aligns with the goal of speeding up development of any
DotStar controllers people are programming.